# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticket Booking System API Documentation",
        "description": "API documentation for the Ticket Booking System backend.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Ticket Booking System API Server"
        }
    ],
    "paths": {
...
        "/payments/initiate/{booking}": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Initiate a payment for a booking",
                "operationId": "a7e1212ce49ef8a3f063ecb8899eb7b5",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "description": "ID of the booking to initiate payment for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment initiated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment initiated"
                                        },
                                        "payment_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "transaction_id": {
                                            "type": "string",
                                            "example": "TRX_abcdef12345"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 150
                                        },
                                        "redirect_url": {
                                            "type": "string",
                                            "example": "http://example.com/sslcommerz-payment-gateway?trxid=TRX_abcdef12345"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/payments/success": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Callback URL for successful payments",
                "operationId": "2133cc54a180a1df31fab3b426ead038",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tran_id"
                                ],
                                "properties": {
                                    "tran_id": {
                                        "type": "string",
                                        "example": "TRX_abcdef12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            }
        },
        "/payments/fail": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Callback URL for failed payments",
                "operationId": "56d13fc8cc1089c784f6b299bbaf3245",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tran_id"
                                ],
                                "properties": {
                                    "tran_id": {
                                        "type": "string",
                                        "example": "TRX_abcdef12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            }
        },
        "/payments/cancel": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Callback URL for cancelled payments",
                "operationId": "c23ba32ea1cfa5ef4348b781753c58ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tran_id"
                                ],
                                "properties": {
                                    "tran_id": {
                                        "type": "string",
                                        "example": "TRX_abcdef12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment cancelled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            }
        },
        "/payments/ipn": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Callback URL for Instant Payment Notification (IPN)",
                "operationId": "f7827da06d9216c802f0edb7c57100c9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tran_id"
                                ],
                                "properties": {
                                    "tran_id": {
                                        "type": "string",
                                        "example": "TRX_abcdef12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "IPN received and processed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "IPN received and processed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Payment": {
                "title": "Payment",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ValidationError"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "int64",
                                "example": 1
                            },
                            "booking_id": {
                                "type": "integer",
                                "format": "int64",
                                "example": 1
                            },
                            "transaction_id": {
                                "type": "string",
                                "example": "TRX_123456789"
                            },
                            "amount": {
                                "type": "number",
                                "format": "float",
                                "example": 150
                            },
                            "currency": {
                                "type": "string",
                                "example": "BDT"
                            },
                            "status": {
                                "type": "string",
                                "example": "pending"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01T00:00:00.000000Z"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01T00:00:00.000000Z"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "ValidationError": {
                "title": "Validation Error",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}